refactor(analysis-types): split api surface DTOs
